Further refactor location groups for reuse
